[cloud-workers] add a retryAt field to do exponential backoffs
Summary:
Add exponetial backoff and cleaned up some functions in the worker.
We now have a `retryAt`. Didn't use the `ExponentialBackoff` class because
that needs to stay in memory the whole time and isn't designed to be
backed by a retryAt field. Just implemented 2^x backoff.

Test Plan: Manual. Threw a `RetryableError`

Reviewers: juan, spang, halla

Reviewed By: halla

Differential Revision: https://phab.nylas.com/D4384