[client-app] record error location to Mixpanel
Summary:
When the client reports an auth error to mixpanel, it could be because the
cloud encountered an error and then reported that back to the client,
which then reported it to mixpanel.

Before we had no way to know where the error came from.

This may help us diagnose errors that come from the client (due to local
ISPs and such) and those that come from the cloud (an AWS environment)

Test Plan: manual

Reviewers: halla, spang, juan

Reviewed By: juan

Differential Revision: https://phab.nylas.com/D4463