[client-app] only poll for identity every hour instead of every 10 min
Summary:
We used to poll for the identity every 10 minutes (which create quited a
bit of load on the billing site). The reason we polled was to get updates
for feature usage. If you used up all of your usage on 1 computer, the
  other needed to get updated eventually.

I don't think this needs to be every 10 minutes. Futhermore, when you make
a request against feature usage right now, it'll optimistcally refresh

Test Plan: manual

Reviewers: juan, spang, halla, khamidou

Reviewed By: khamidou

Differential Revision: https://phab.nylas.com/D4486