[client-sync] Keep track of when SendTaskRunners expire
Summary:
Although we check if the account needs a new token before each send,
we failed to consider that the token in the account might not match
the token used in the SendTaskRunner. This could happen if the token
was refreshed somewhere outside the SendTaskManager (e.g. the sync
loop). This diff adds an `expiresAtInMs` field to the SendTaskRunner
so that the SendTaskManager can determine whether it needs an
updated SendTaskRunner or not.

Test Plan: manual

Reviewers: juan, evan

Reviewed By: juan, evan

Differential Revision: https://phab.nylas.com/D4515